ECU-9177 Unify model Project into ProjectKey
